// src/app/interfaces/NivelConocimiento.ts
export interface DataNivelConocmiento {
    conocePresidenteMunicipal: string;
    nombrePresidenteMunicipal: string;
    resultadoNombrePresidenteMunicipal: string;
    idPartidoPoliticoPresidenteMunicipal: string;
    conoceGobernadorEstado: string;
    nombreGobernadorEstado: string;
    resultadoGobernadorEstado: string;
    idPartidoPoliticoGobernadorEstado: string;
    conocePresidenteMexico: string;
    nombrePresidenteMexico: string;
    resultadoNombrePresidenteMexico: string;
    idPartidoPoliticoPresidenteMexico: string;
    idEleccionesEstado: string;
    nombreCandidatoPanEstadoMexico: string;
    resultadoNombreCandidatoPanEstadoMexico: string;
    nombreCandidatoPriEstadoMexico: string;
    resultadoNombreCandidatoPriEstadoMexico: string;
    nombreCandidatoPrdEstadoMexico: string;
    resultadoNombreCandidatoPrdEstadoMexico: string;
    nombreCandidatoMorenaEstadoMexico: string;
    resultadoNombreCandidatoMorenaEstadoMexico: string;
}